+++ tests/fixtures_data/daily-k.pkl
